Create home_inventory_manager.java
This is a GUI Application to store the details of one's belongings. The details include item name, item number, location, purchase price and date, picture and name of store from where purchased. The details of the belongings are stored and can be retrieved and modified whenever required.  They can also be deleted whenever required.